Refactored code to correctly reach endpoints
